Rename from `protop2g` to `pagure-exporter`
