+++ .pre-commit-config.yaml
@@ -28,8 +28,6 @@ repos:
     rev: 5.12.0
     hooks:
     -   id: isort
-        name: isort (python)
-    -   id: isort
         name: isort (cython)
         types: [cython]
     -   id: isort
