+++ pagure_exporter/view/stat.py
@@ -1,5 +1,5 @@
 """
-protop2g
+Pagure Exporter
 Copyright (C) 2022-2023 Akashdeep Dhar
 
 This program is free software: you can redistribute it and/or modify
@@ -25,9 +25,9 @@ of Red Hat, Inc.
 import sys
 from time import localtime, strftime
 
-from protop2g.conf import standard
-from protop2g.view.dcrt import failure, general, section, success
-from protop2g.work.stat import DestData, SrceData
+from pagure_exporter.conf import standard
+from pagure_exporter.view.dcrt import failure, general, section, success
+from pagure_exporter.work.stat import DestData, SrceData
 
 
 def showstat():
