+++ pagure_exporter/work/keep.py
@@ -1,5 +1,5 @@
 """
-protop2g
+Pagure Exporter
 Copyright (C) 2022-2023 Akashdeep Dhar
 
 This program is free software: you can redistribute it and/or modify
@@ -22,7 +22,7 @@ of Red Hat, Inc.
 """
 
 
-from protop2g.conf import standard
+from pagure_exporter.conf import standard
 
 
 def storeinf(srce, dest, pkey, gkey, fusr, tusr):
