+++ tox.ini
@@ -20,10 +20,10 @@ deps =
 commands_pre =
     poetry install --all-extras
 commands =
-  poetry run pytest -o 'addopts=--cov-config .coveragerc --cov=protop2g --cov-report term-missing --cov-report xml --cov-report html' test/ {posargs} -vvv
+    poetry run pytest -o 'addopts=--cov-config .coveragerc --cov=pagure_exporter --cov-report term-missing --cov-report xml --cov-report html' test/ {posargs} -vvv
 
 # To let your instances of `print` commands appear on stdout in the tests for debugging, use `pytest -s` on the above command.
 
 [testenv:cleaning]
 commands =
-  poetry run ruff check protop2g/ test/
+    poetry run ruff check pagure_exporter/ test/
